requirements.txt
# =============================================================================
# Naboom Community Platform - Production Requirements
# =============================================================================
# Django 5.2 + Wagtail 7.1.1 Production Dependencies
# =============================================================================

# Core Django Framework
Django==5.2.7
Wagtail==7.1.1

# Database
psycopg2-binary==2.9.9
django-environ==0.11.2

# Caching & Sessions
redis==5.0.1
django-redis==5.4.0

# Web Server
gunicorn==21.2.0
daphne==4.0.0

# Real-time Features
channels==4.0.0
channels-redis==4.2.0

# Background Tasks
celery==5.3.4
django-celery-beat==2.5.0

# Storage & Media
django-storages==1.14.2
boto3==1.34.34
Pillow==10.2.0

# API & CORS
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Security
django-ratelimit==4.1.0
django-csp==3.7

# Monitoring & Performance
django-silk==5.0.4
django-debug-toolbar==4.2.0

# Utilities
python-decouple==3.8
whitenoise==6.6.0
django-extensions==3.2.3

# Development Tools (optional)
pytest==7.4.4
pytest-django==4.7.0
coverage==7.4.0
black==23.12.1
flake8==7.0.0
isort==5.13.2
